Refactor: Moved Response objects to a seperate subfolder
