+++ lib/Builder.ts
@@ -2,8 +2,8 @@ import {Model} from "./Model";
 import {FilterSpec} from "./FilterSpec";
 import {SortSpec} from "./SortSpec";
 import {AxiosResponse, AxiosError, AxiosInstance} from "axios";
-import {PluralResponse} from "./PluralResponse";
-import {SingularResponse} from "./SingularResponse";
+import {PluralResponse} from "./response/PluralResponse";
+import {SingularResponse} from "./response/SingularResponse";
 import {Promise} from 'es6-promise';
 import axios from 'axios';
 import {Option} from "./Option";
