+++ lib/response/Response.ts
@@ -1,11 +1,11 @@
-import {JsonApiResponseBody} from "./JsonApiResponseBody";
-import {JsonApiDoc} from "./JsonApiDoc";
-import {Model} from "./Model";
-import {ToManyRelation} from "./ToManyRelation";
-import {JsonApiStub} from "./JsonApiStub";
-import {Relation} from "./Relation";
-import {ToOneRelation} from "./ToOneRelation";
-import {Map} from "./util/Map";
+import {JsonApiResponseBody} from "../JsonApiResponseBody";
+import {JsonApiDoc} from "../JsonApiDoc";
+import {Model} from "../Model";
+import {ToManyRelation} from "../ToManyRelation";
+import {JsonApiStub} from "../JsonApiStub";
+import {Relation} from "../Relation";
+import {ToOneRelation} from "../ToOneRelation";
+import {Map} from "../util/Map";
 
 export abstract class Response
 {
